# services/news-feed-engine/api/openapi.yaml
openapi: 3.1.0
info:
  title: ElevatedIQ News Feed Engine API
  description: |
    AI-powered news aggregation platform API for multi-platform content ingestion,
    analysis, and video summary generation. Supports white-label multi-tenant deployment.
  version: 1.0.0
  contact:
    name: ElevatedIQ Platform Team
    email: platform@elevatediq.ai
  license:
    name: Proprietary
servers:
  - url: https://api.elevatediq.ai/news-feed/v1
    description: Production
  - url: https://staging-api.elevatediq.ai/news-feed/v1
    description: Staging
  - url: http://127.0.0.1:8096/api/v1
    description: Local Development
...
paths:
...
  /videos:
    get:
      tags: [Videos]
      summary: List video summaries
      operationId: listVideos
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - $ref: '#/components/parameters/category'
      responses:
        '200':
          description: Paginated list of videos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoListResponse'
  /videos/{id}:
    get:
      tags: [Videos]
      summary: Get video by ID
      operationId: getVideo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/{id}/transcript:
    get:
      tags: [Videos]
      summary: Get video transcript
      operationId: getVideoTranscript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          schema:
            type: string
            enum: [text, srt, vtt, json]
            default: text
      responses:
        '200':
          description: Video transcript
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptResponse'
            text/plain:
              schema:
                type: string
...
  /admin/videos/generate:
    post:
      tags: [Admin]
      summary: Generate video summary
      operationId: generateVideo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoGenerationRequest'
      responses:
        '202':
          description: Video generation job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/whitelabel:
    get:
      tags: [Admin]
      summary: Get white-label configuration
      operationId: getWhitelabelConfig
      security:
        - bearerAuth: []
      responses:
        '200':
          description: White-label configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhitelabelConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [Admin]
      summary: Update white-label configuration
      operationId: updateWhitelabelConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhitelabelConfig'
      responses:
        '200':
          description: White-label configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    category:
      name: category
      in: query
      schema:
        type: string
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_more:
          type: boolean
...
    Video:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content_id:
          type: string
          format: uuid
        title:
          type: string
        video_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
        duration:
          type: integer
          description: Duration in seconds
        status:
          type: string
          enum: [pending, processing, completed, failed]
        view_count:
          type: integer
        generated_at:
          type: string
          format: date-time
    WhitelabelConfig:
      type: object
      properties:
        tenant_id:
          type: string
        display_name:
          type: string
        logo_url:
          type: string
          format: uri
        favicon_url:
          type: string
          format: uri
        primary_color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
        secondary_color:
          type: string
        accent_color:
          type: string
        font_family:
          type: string
        custom_domain:
          type: string
        enabled_platforms:
          type: array
          items:
            type: string
        enabled_categories:
          type: array
          items:
            type: string
        settings:
          type: object
...
    VideoGenerationRequest:
      type: object
      required:
        - content_id
      properties:
        content_id:
          type: string
          format: uuid
        template_id:
          type: string
        voice_id:
          type: string
        avatar_id:
          type: string
        resolution:
          type: string
          enum: [720p, 1080p, 4k]
        priority:
          type: integer
          minimum: 1
          maximum: 10
...
    VideoListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Video'
        pagination:
          $ref: '#/components/schemas/Pagination'
    VideoResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Video'
    TranscriptResponse:
      type: object
      properties:
        video_id:
          type: string
        format:
          type: string
        transcript:
          type: string
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string

    # AI Schemas
...